Move cli options to generate_config
