+++ .coveragerc
@@ -8,6 +8,8 @@ omit =
     *_mako
 
 [report]
+show_missing = True
+skip_covered = True
 exclude_lines =
     # Have to re-enable the standard pragma
     \#\s*pragma: no cover
