+++ git_code_debt/options.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
-import six
+from git_code_debt.generate_config import DEFAULT_GENERATE_CONFIG_FILENAME
 
 
 # TODO: Take the color processing code from pre-commit and package it
@@ -14,33 +14,8 @@ def add_color(argparser):
     )
 
 
-def add_database(argparser):
+def add_generate_config_filename(argparser):
     argparser.add_argument(
-        'database',
-        help='Database filename (usually end with .db).',
-    )
-
-
-def add_metric_package_names(argparser):
-    argparser.add_argument(
-        'metric_package_names',
-        type=six.text_type,
-        nargs='*',
-        help='Metric Package Names (such as foo.metrics bar.metrics).',
-    )
-
-
-def add_repo(argparser):
-    argparser.add_argument(
-        'repo',
-        help='Git url to a repository to generate metrics from.',
-    )
-
-
-def add_skip_default_metrics(argparser):
-    argparser.add_argument(
-        '--skip-default-metrics',
-        default=False,
-        action='store_true',
-        help='Whether to skip default metrics.',
+        '-C', '--config-filename', default=DEFAULT_GENERATE_CONFIG_FILENAME,
+        help='Path to generate config.',
     )
