+++ setup.py
@@ -44,7 +44,6 @@ setup(
 
     entry_points={
         'console_scripts': [
-            'git-code-debt-create-tables = git_code_debt.create_tables:main',
             'git-code-debt-generate = git_code_debt.generate:main',
             'git-code-debt-list-metrics = git_code_debt.list_metrics:main',
             'git-code-debt-server = git_code_debt.server.app:main',
