+++ tests/discovery_test.py
@@ -64,7 +64,7 @@ def test_get_modules_no_modules():
 
 
 def test_get_modules_some_modules():
-    module_names = ['git_code_debt.metrics', 'git_code_debt.create_tables']
+    module_names = ['git_code_debt.metrics', 'git_code_debt.generate']
     ret = get_modules(module_names)
     assert len(ret) == 2
     assert [mod.__name__ for mod in ret] == module_names
