+++ tests/logic_test.py
@@ -1,8 +1,8 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
-from git_code_debt.create_tables import get_metric_ids
 from git_code_debt.discovery import get_metric_parsers
+from git_code_debt.generate import get_metric_ids
 from git_code_debt.logic import get_metric_mapping
 from git_code_debt.logic import get_metric_values
 from git_code_debt.logic import get_previous_sha
