+++ tox.ini
@@ -9,7 +9,7 @@ usedevelop = True
 commands =
     coverage erase
     coverage run -m pytest {posargs:tests}
-    coverage report --show-missing --fail-under 100
+    coverage report --fail-under 100
     pre-commit run --all-files
 
 [testenv:venv]
